extconf: Link with math when building the native extension
When we used the upstream gem we encountered the following error:

LoadError: /var/sites/skroutz_cap/shared/bundle/ruby/2.3.0/gems/phashion-1.2.0/lib/phashion_ext.so: undefined symbol: _ZGVbN2v_cos

The symbol cos is provided by `math.h` which is not linked when building
the native extension.

This commit adds the `-lm` flag in the `$LIBS` in order to link with
math.