Adjust thresholds in tests
This commit adjusts the thresholds in two tests regarding the mexican
hat calculation functionality. It increases the thressholds from the
value of `0.028` to `0.033`.
The same two tests are failing for Debian Stretch with ruby 2.3.3
and for Debian Buster with ruby 2.5.5. There is also a github issue
regarding these same tests for Ubuntu 16.04[1]
It seems that this issue is a floating point issue but somewhat
related to the OS, because linux trusty specs[2] pass whereas the specs for
linux xenial[3] do not and the code has not changed.
See also here[4], where the thresholds for these same tests have
been modified by the original authors without justifying the change.

[1]: https://github.com/westonplatter/phashion/issues/78
[2]: https://travis-ci.org/github/MatzFan/phashion/jobs/531359477
[3]: https://travis-ci.org/github/MatzFan/phashion/jobs/531367883
[4]: https://github.com/skroutz/phashion/commit/b6f3a77579095d367ed9998f9eed8b8f36a9ef40